Shared ViewModel Across Fragments 5. Use the ViewModel to update the UI
In this task, you will use the shared view model you created to update the app's UI. The main difference in the implementation of a shared view model is the way we access it from the UI controllers. You will use the activity instance instead of the fragment instance, and you will see how to do this in the coming sections.

That means the view model can be shared across fragments. Each fragment could access the view model to check on some detail of the order or update some data in the view model.